fix #1474 move bag encryption plugins into separate library (#1499)
* fix #1474 move bag encryption plugins into separate library

This setup is required by the class_loader, which is part of the pluginlib, see  http://wiki.ros.org/class_loader#Caution_of_Linking_Directly_Against_Plugin_Libraries

* fix unit tests of aes_encryptor
